Move stix_sdo to core and have Entity inherit from it.
