+++ yeti/core/entities/malware.py
@@ -1,10 +1,6 @@
 """Detail Yeti's Malware object structure."""
 
-from marshmallow import fields
-
-from yeti.core.errors import ValidationError
-from .stix_entity import Entity
-
+from .entity import Entity
 
 class Malware(Entity):
     """Malware Yeti object.
